Cleaned all existing code (deleteing redundant or outdated comments/functions) and added fully functional delete product.
