+++ db/index.js
@@ -14,11 +14,4 @@ module.exports = {
   ...require('./orders'),
   ...require('./order_products'),
   ...require('./reviews'),
-
-  //For some reason this doesn't work but I'm keeping itanywa to troubleshoot later
-  // ...require('./init_categories'),
-  // ...require('./init_orders'),
-  // ...require('./init_products'),
-  // ...require('./init_reviews'),
-  // ...require('./init_users'),
 };
