+++ src/api/index.js
@@ -1,40 +1,6 @@
 import axios from 'axios';
 
 
-// export async function getSomething() {
-//   try {
-//     const { data } = await axios.get('/api');
-//     return data;
-//   } catch (error) {
-//     throw error;
-//   }
-// }
-
 const BASE_URL = '/api';
 
-export async function getAllUsers() {
-  try {
-    const { data } = await axios.get(`${BASE_URL}/users`);
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
-
-export async function getAllProducts() {
-  try {
-    const { data } = await axios.get(`${BASE_URL}/products`)
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
-
-export async function deleteProduct(productId) {
-  try {
-    const { data } = await axios.get(`${BASE_URL}/products/${productId}`);
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
+//Maybe this should just be where we import/export all the api functions for use in componets?
