+++ src/components/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 
 import Content, { }
   from './Content.js';
