+++ src/components/NewProductForm.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const NewProductForm = () => {
+
+    return (
+        <h1>Create New Product From</h1>
+    )
+};
+
+export default NewProductForm;
